+++ sdk/keyvault/azure-security-keyvault-keys/inc/azure/keyvault/keys/key_operation.hpp
@@ -9,6 +9,8 @@
 
 #pragma once
 
+#include "azure/keyvault/keys/dll_import_export.hpp"
+
 #include <string>
 
 namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
@@ -27,7 +29,23 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
      *
      * @param operation The operation for the key as string.
      */
-    KeyOperation(std::string const& operation) : m_operation(operation) {}
+    explicit KeyOperation(std::string operation) : m_operation(std::move(operation)) {}
+
+    /**
+     * @brief Construct a default Key operation.
+     *
+     */
+    KeyOperation() = default;
+
+    /**
+     * @brief Enables using the equal operator for key operations.
+     *
+     * @param other A key operation to be compared.
+     */
+    bool operator==(const KeyOperation& other) const noexcept
+    {
+      return m_operation == other.m_operation;
+    }
 
     /**
      * @brief Returns the fully qualified type name of this instance.
@@ -42,7 +60,7 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
      *
      * @return Encrypt KeyOperation.
      */
-    static KeyOperation Encrypt() { return KeyOperation("encrypt"); }
+    AZ_SECURITY_KEYVAULT_KEYS_DLLEXPORT static const KeyOperation Encrypt;
 
     /**
      * @brief The key can be used to decrypt with the #Decrypt(EncryptionAlgorithm, Byte[],
@@ -50,7 +68,7 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
      *
      * @return Decrypt KeyOperation.
      */
-    static KeyOperation Decrypt() { return KeyOperation("decrypt"); }
+    AZ_SECURITY_KEYVAULT_KEYS_DLLEXPORT static const KeyOperation Decrypt;
 
     /**
      * @brief The key can be used to sign with the Sign(SignatureAlgorithm, Byte[],
@@ -58,7 +76,7 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
      *
      * @return Sign KeyOperation.
      */
-    static KeyOperation Sign() { return KeyOperation("sign"); }
+    AZ_SECURITY_KEYVAULT_KEYS_DLLEXPORT static const KeyOperation Sign;
 
     /**
      * @brief The key can be used to verify with the Verify(SignatureAlgorithm, Byte[], Byte[],
@@ -66,7 +84,7 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
      *
      * @return Verify KeyOperation.
      */
-    static KeyOperation Verify() { return KeyOperation("verify"); }
+    AZ_SECURITY_KEYVAULT_KEYS_DLLEXPORT static const KeyOperation Verify;
 
     /**
      * @brief The key can be used to wrap another key with the WrapKey(KeyWrapAlgorithm, Byte[],
@@ -74,7 +92,7 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
      *
      * @return WrapKey KeyOperation.
      */
-    static KeyOperation WrapKey() { return KeyOperation("wrapKey"); }
+    AZ_SECURITY_KEYVAULT_KEYS_DLLEXPORT static const KeyOperation WrapKey;
 
     /**
      * @brief The key can be used to unwrap another key with the UnwrapKey(KeyWrapAlgorithm, Byte[],
@@ -82,7 +100,7 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
      *
      * @return UnwrapKey KeyOperation.
      */
-    static KeyOperation UnwrapKey() { return KeyOperation("unwrapKey"); }
+    AZ_SECURITY_KEYVAULT_KEYS_DLLEXPORT static const KeyOperation UnwrapKey;
 
     /**
      * @brief The key can be imported during creation using the ImportKey(ImportKeyOptions,
@@ -90,7 +108,7 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
      *
      * @return Import KeyOperation.
      */
-    static KeyOperation Import() { return KeyOperation("import"); }
+    AZ_SECURITY_KEYVAULT_KEYS_DLLEXPORT static const KeyOperation Import;
   };
 
 }}}} // namespace Azure::Security::KeyVault::Keys