+++ sdk/keyvault/azure-security-keyvault-keys/test/samples/sample1-hello-world/sample1_hello_world.cpp
@@ -49,7 +49,7 @@ int main()
 
     KeyVaultKey cloudRsaKey = keyClient.GetKey(rsaKeyName).Value;
     std::cout << "Key is returned with name " << cloudRsaKey.Name() << " and type "
-              << KeyType::KeyTypeToString(cloudRsaKey.GetKeyType()) << std::endl;
+              << cloudRsaKey.GetKeyType().ToString() << std::endl;
 
     cloudRsaKey.Properties.ExpiresOn
         = cloudRsaKey.Properties.ExpiresOn.Value() + std::chrono::hours(24 * 365);