+++ sdk/keyvault/azure-security-keyvault-keys/test/samples/sample3-get-keys/sample3_get_keys.cpp
@@ -66,7 +66,7 @@ int main()
         }
         auto keyWithType = keyClient.GetKey(key.Name).Value;
         std::cout << "Key is returned with name: " << keyWithType.Name()
-                  << " and type: " << KeyType::KeyTypeToString(keyWithType.GetKeyType())
+                  << " and type: " << keyWithType.GetKeyType().ToString()
                   << std::endl;
       }