+++ sdk/keyvault/azure-security-keyvault-keys/test/ut/key_client_delete_test_live.cpp
@@ -315,8 +315,6 @@ TEST_F(KeyVaultClientTest, GetDeletedKey)
     EXPECT_FALSE(deletedKey.RecoveryId.empty());
     EXPECT_EQ(deletedKey.Name(), keyName);
     auto expectedType = Azure::Security::KeyVault::Keys::JsonWebKeyType::Ec;
-    EXPECT_EQ(
-        Azure::Security::KeyVault::Keys::KeyType::KeyTypeToString(expectedType),
-        Azure::Security::KeyVault::Keys::KeyType::KeyTypeToString(deletedKey.Key.KeyType));
+    EXPECT_EQ(expectedType, deletedKey.Key.KeyType);
   }
 }